March development changes
New 'pin' command supported. Lets the RPi control ON/OFF status of individual GPIO pins. Really a dev/test/debug feature. Makes it easier to turn pins on/off and then measure the behaviour of the motorcontroller circuit.
Removed some unused code.
Reduced the number of components imported by the program. Runtime memory is lower on CircuitPython 8.2, so some housekeeping was wise to allow space for further development in the future.