March & April development updates
Multiple development changes during March and April included in this.
I'm currently testing this copy across RPi3, RPi4 and RPi5 platforms.
If you pull this version you must pull the entire branch, there are related changes in circuitpython/code.py which need to be used too.

This introduces the new pilomargpio include file as well. This is to support different GPIO handlers between RPi3/4 and RPi5.